Included Mapview in MapHistoryFragment
