+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,17 @@
         >
         <meta-data android:name="AA_DB_NAME" android:value="sensor_records" />
         <meta-data android:name="AA_DB_VERSION" android:value="1" />
+        <meta-data
+            android:name="com.google.android.gms.version"
+            android:value="@integer/google_play_services_version" />
+        <meta-data
+            android:name="com.google.android.geo.API_KEY"
+            android:value="AIzaSyDkRzwTszDaQE_BgNLcuUdQmv7cV76eF1A"/>
+
+        <uses-feature
+            android:glEsVersion="0x00020000"
+            android:required="true"/>
+
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -26,7 +37,16 @@
         </activity>
     </application>
 
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
+    <permission
+        android:name="your.package.name.permission.MAPS_RECEIVE"
+        android:protectionLevel="signature" />
+    <uses-permission android:name="your.package.name.permission.MAPS_RECEIVE"/>
+
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
+    <uses-permission android:name="com.google.android.providers.gsf.permission.READ_GSERVICES" />
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
 
 </manifest>
